Split DataSourceStream contents into separate files
No code changes have been made, just rearranging content and tidying #includes.

Unexpected conflicts occurred with libmain (see below) because of module re-ordering. The simplest solution I could find was to put LIBSMING before LIBMAIN in the makefiles.

```
out/build/libmain2.a(time.o): In function `millis':
(.irom0.text+0xb4): undefined reference to `sntp_set_update_delay'
out/build/libmain2.a(time.o): In function `millis':
(.irom0.text+0xbd): undefined reference to `sntp_set_update_delay'
out/build/libmain2.a(time.o): In function `micros':
(.irom0.text+0xe0): undefined reference to `sntp_set_daylight'
out/build/libmain2.a(time.o): In function `micros':
(.irom0.text+0xe4): undefined reference to `sntp_set_timetype'
out/build/libmain2.a(time.o): In function `updateTime':
(.irom0.text+0xe8): undefined reference to `sntp_set_receive_time_size'
out/build/libmain2.a(time.o): In function `updateTime':
(.irom0.text+0x134): undefined reference to `sntp_set_daylight'
out/build/libmain2.a(time.o): In function `configTime':
(.irom0.text+0x13c): undefined reference to `sntp_set_timetype'
out/build/libmain2.a(time.o): In function `configTime':
(.irom0.text+0x169): undefined reference to `sntp_set_receive_time_size'
out/build/libmain2.a(time.o): In function `configTime':
(.irom0.text+0x1a8): undefined reference to `sntp_get_timetype'
out/build/libmain2.a(time.o): In function `configTime':
(.irom0.text+0x1bb): undefined reference to `sntp_get_timetype'
```
