+++ Sming/SmingCore/Data/Stream/ChunkedStream.cpp
@@ -1,3 +1,13 @@
+/****
+ * Sming Framework Project - Open Source framework for high efficiency native ESP8266 development.
+ * Created 2015 by Skurydin Alexey
+ * http://github.com/SmingHub/Sming
+ * All files of the Sming Core are provided under the LGPL v3 license.
+ *
+ * @author Slavey Karadzhov <slaff@attachix.com>
+ *
+ ****/
+
 #include "ChunkedStream.h"
 
 ChunkedStream::ChunkedStream(ReadWriteStream* stream, size_t resultSize /* = 512 */)
