+++ Sming/SmingCore/Data/Stream/ChunkedStream.h
@@ -8,8 +8,8 @@
  *
  ****/
 
-#ifndef _SMING_CORE_DATA_CHUNKEDSTREAM_H_
-#define _SMING_CORE_DATA_CHUNKEDSTREAM_H_
+#ifndef _SMING_CORE_DATA_CHUNKED_STREAM_H_
+#define _SMING_CORE_DATA_CHUNKED_STREAM_H_
 
 #include "../StreamTransformer.h"
 
@@ -38,4 +38,4 @@ public:
 };
 
 /** @} */
-#endif /* _SMING_CORE_DATA_CHUNKEDSTREAM_H_ */
+#endif /* _SMING_CORE_DATA_CHUNKED_STREAM_H_ */
