+++ Sming/SmingCore/Data/Stream/DataSourceStream.h
@@ -5,17 +5,12 @@
  * All files of the Sming Core are provided under the LGPL v3 license.
  ****/
 
-#ifndef _SMING_CORE_DATA_DATASTREAM_H_
-#define _SMING_CORE_DATA_DATASTREAM_H_
+#ifndef _SMING_CORE_DATA_DATA_SOURCE_STREAM_H_
+#define _SMING_CORE_DATA_DATA_SOURCE_STREAM_H_
 
 #include <user_config.h>
-#include "../SmingCore/FileSystem.h"
-#include "../Libraries/ArduinoJson/include/ArduinoJson.h"
-#include "../Wiring/WString.h"
-#include "../Wiring/WHashMap.h"
-#include "../Wiring/Stream.h"
-
-#define TEMPLATE_MAX_VAR_NAME_LEN 16
+#include "Stream.h"
+#include "WString.h"
 
 /** @brief  Data stream type
  *  @ingroup constants
@@ -35,13 +30,6 @@ enum StreamType {
  *  @brief      Data stream classes
  *  @{
 */
-
-/** @brief  Template variable (hash map) class
- *  @see    Wiring HashMap
- */
-class TemplateVariables : public HashMap<String, String>
-{
-};
 
 ///Base class for data source stream
 class IDataSourceStream : public Stream
@@ -78,12 +66,12 @@ public:
 	virtual int peek();
 
 	/** @brief  Move read cursor
-	 *  @param  len Position within stream to move cursor to
+	 *  @param  len Relative cursor adjustment
 	 *  @retval bool True on success.
 	 */
 	virtual bool seek(int len) = 0;
 
-	/** @brief  Check if stream is finished
+	/** @brief  Check if all data has been read
      *  @retval bool True on success.
      */
 	virtual bool isFinished() = 0;
@@ -126,334 +114,5 @@ public:
 	}
 };
 
-class ReadWriteStream : public IDataSourceStream
-{
-public:
-	virtual ~ReadWriteStream()
-	{
-	}
-
-	virtual size_t write(uint8_t charToWrite) = 0;
-
-	/** @brief  Write chars to stream
-     *  @param  buffer Pointer to buffer to write to the stream
-     *  @param  size Quantity of chars to write
-     *  @retval size_t Quantity of chars written to stream
-     */
-	virtual size_t write(const uint8_t* buffer, size_t size) = 0;
-
-	//Use base class documentation
-	virtual uint16_t readMemoryBlock(char* data, int bufSize) = 0;
-};
-
-/// Memory data stream class
-class MemoryDataStream : public ReadWriteStream
-{
-public:
-	/** @brief Memory data stream base class
-    */
-	MemoryDataStream();
-	virtual ~MemoryDataStream();
-
-	//Use base class documentation
-	virtual StreamType getStreamType()
-	{
-		return eSST_Memory;
-	}
-
-	/** @brief  Get a pointer to the current position
-	 *  @retval "const char*" Pointer to current cursor position within the data stream
-	 */
-	const char* getStreamPointer()
-	{
-		return pos;
-	}
-
-	/**
-	 * @brief Return the total length of the stream
-	 * @retval int -1 is returned when the size cannot be determined
-	*/
-	int available();
-
-	/** @brief  Write a single char to stream
-     *  @param  charToWrite Char to write to the stream
-     *  @retval size_t Quantity of chars written to stream (always 1)
-     */
-	virtual size_t write(uint8_t charToWrite);
-
-	/** @brief  Write chars to stream
-     *  @param  buffer Pointer to buffer to write to the stream
-     *  @param  size Quantity of chars to write
-     *  @retval size_t Quantity of chars written to stream
-     */
-	virtual size_t write(const uint8_t* buffer, size_t size);
-
-	//Use base class documentation
-	virtual uint16_t readMemoryBlock(char* data, int bufSize);
-
-	//Use base class documentation
-	virtual bool seek(int len);
-
-	//Use base class documentation
-	virtual bool isFinished();
-
-private:
-	char* buf;
-	char* pos;
-	int size;
-	int capacity;
-};
-
-/// File stream class
-class FileStream : public ReadWriteStream
-{
-public:
-	/** @brief  Create a file stream
-     *  @param  fileName Name of file to open
-     */
-	FileStream();
-	FileStream(const String& fileName);
-	virtual ~FileStream();
-
-	virtual bool attach(const String& fileName, FileOpenFlags openFlags);
-	//Use base class documentation
-	virtual StreamType getStreamType()
-	{
-		return eSST_File;
-	}
-
-	virtual size_t write(uint8_t charToWrite);
-	virtual size_t write(const uint8_t* buffer, size_t size);
-
-	//Use base class documentation
-	virtual uint16_t readMemoryBlock(char* data, int bufSize);
-
-	//Use base class documentation
-	virtual bool seek(int len);
-
-	//Use base class documentation
-	virtual bool isFinished();
-
-	String fileName(); ///< Filename of file stream is attached to
-	bool fileExist();  ///< True if file exists
-
-	/** @brief  Get the offset of cursor from beginning of data
-     *  @retval int Cursor offset
-     */
-	inline int getPos()
-	{
-		return pos;
-	}
-
-	/**
-	 * @brief Return the total length of the stream
-	 * @retval int -1 is returned when the size cannot be determined
-	 */
-	int available()
-	{
-		return size;
-	}
-
-	virtual String id();
-
-private:
-	file_t handle;
-	int pos;
-	int size;
-};
 /** @} */
-
-/** @brief  Template file stream expand state
- *  @ingroup constants
- *  @{
- */
-enum TemplateExpandState {
-	eTES_Wait,		///< Template expand state wait
-	eTES_Found,		///< Template expand state found
-	eTES_StartVar,  ///< Template expand state start variable
-	eTES_SendingVar ///< Template expand state sending variable
-};
-/** @} */
-
-/** @addtogroup stream
- *  @{
- */
-
-/// Template file stream class
-class TemplateFileStream : public FileStream
-{
-public:
-	/** @brief Create a template file stream
-     *  @param  templateFileName Template filename
-     */
-	TemplateFileStream(const String& templateFileName);
-	virtual ~TemplateFileStream();
-
-	//Use base class documentation
-	virtual StreamType getStreamType()
-	{
-		return eSST_TemplateFile;
-	}
-
-	//Use base class documentation
-	virtual uint16_t readMemoryBlock(char* data, int bufSize);
-
-	//Use base class documentation
-	virtual bool seek(int len);
-
-	/** @brief  Set value of a variable in the template file
-     *  @param  name Name of variable
-     *  @param  value Value to assign to the variable
-     *  @note   Sets and existing variable or adds a new variable if variable does not already exist
-     */
-	void setVar(String name, String value);
-
-	/** @brief  Set multiple variables in the template file
-     *  @param  vars Template Variables
-     */
-	void setVars(const TemplateVariables& vars);
-
-	/** @brief  Get the template variables
-     *  @retval TemplateVariables Reference to the template variables
-     */
-	inline TemplateVariables& variables()
-	{
-		return templateData;
-	}
-
-	/**
-	 * @brief Return the total length of the stream
-	 * @retval int -1 is returned when the size cannot be determined
-	 */
-	int available()
-	{
-		return -1;
-	}
-
-private:
-	TemplateVariables templateData;
-	TemplateExpandState state;
-	String varName;
-	int skipBlockSize = 0;
-	int varDataPos = 0;
-	int varWaitSize = 0;
-};
-
-///JSON object stream class
-class JsonObjectStream : public MemoryDataStream
-{
-public:
-	/** @brief  Create a JSON object stream
-    */
-	JsonObjectStream();
-	virtual ~JsonObjectStream();
-
-	//Use base class documentation
-	virtual StreamType getStreamType()
-	{
-		return eSST_JsonObject;
-	}
-
-	/** @brief  Get the JSON root node
-     *  @retval JsonObject Reference to the root node
-     */
-	JsonObject& getRoot();
-
-	//Use base class documentation
-	virtual uint16_t readMemoryBlock(char* data, int bufSize);
-
-	/**
-	 * @brief Return the total length of the stream
-	 * @retval int -1 is returned when the size cannot be determined
-	 */
-	int available();
-
-private:
-	DynamicJsonBuffer buffer;
-	JsonObject& rootNode;
-	bool send;
-};
-
-class EndlessMemoryStream : public ReadWriteStream
-{
-public:
-	virtual ~EndlessMemoryStream();
-
-	//Use base class documentation
-	virtual StreamType getStreamType();
-
-	virtual uint16_t readMemoryBlock(char* data, int bufSize);
-
-	//Use base class documentation
-	virtual bool seek(int len);
-
-	/** @brief  Write a single char to stream
-	 *  @param  charToWrite Char to write to the stream
-	 *  @retval size_t Quantity of chars written to stream (always 1)
-	 */
-	virtual size_t write(uint8_t charToWrite);
-
-	/** @brief  Write chars to stream
-	 *  @param  buffer Pointer to buffer to write to the stream
-	 *  @param  size Quantity of chars to write
-	 *  @retval size_t Quantity of chars written to stream
-	 */
-	virtual size_t write(const uint8_t* buffer, size_t size);
-
-	virtual bool isFinished();
-
-private:
-	MemoryDataStream* stream = NULL;
-};
-
-/**
- * @brief Memory stream that stores limited number of bytes
- * 		  Once the limit is reached the stream will discard incoming bytes on write
- */
-class LimitedMemoryStream : public ReadWriteStream
-{
-public:
-	LimitedMemoryStream(size_t length);
-	virtual ~LimitedMemoryStream();
-
-	//Use base class documentation
-	virtual StreamType getStreamType();
-
-	/**
-	 * @brief Return the total length of the stream
-	 * @retval int -1 is returned when the size cannot be determined
-	 */
-	virtual int available()
-	{
-		return writePos - readPos;
-	}
-
-	virtual uint16_t readMemoryBlock(char* data, int bufSize);
-
-	//Use base class documentation
-	virtual bool seek(int len);
-
-	/** @brief  Write a single char to stream
-	 *  @param  charToWrite Char to write to the stream
-	 *  @retval size_t Quantity of chars written to stream (always 1)
-	 */
-	virtual size_t write(uint8_t charToWrite);
-
-	/** @brief  Write chars to stream
-	 *  @param  buffer Pointer to buffer to write to the stream
-	 *  @param  size Quantity of chars to write
-	 *  @retval size_t Quantity of chars written to stream
-	 */
-	virtual size_t write(const uint8_t* buffer, size_t size);
-
-	virtual bool isFinished();
-
-private:
-	uint8_t* buffer = NULL;
-	size_t writePos = 0;
-	size_t readPos = 0;
-	size_t length = 0;
-};
-
-/** @} */
-#endif /* _SMING_CORE_DATA_DATASTREAM_H_ */
+#endif /* _SMING_CORE_DATA_DATA_SOURCE_STREAM_H_ */
