+++ Sming/SmingCore/Data/Stream/UrlencodedOutputStream.h
@@ -8,11 +8,11 @@
  *
  ****/
 
-#ifndef _SMING_CORE_DATA_UESTREAM_H_
-#define _SMING_CORE_DATA_UESTREAM_H_
+#ifndef _SMING_CORE_DATA_URL_ENCODDED_OUTPUT_STREAM_H_
+#define _SMING_CORE_DATA_URL_ENCODDED_OUTPUT_STREAM_H_
 
-#include "../../Wiring/WHashMap.h"
-#include "DataSourceStream.h"
+#include "MemoryDataStream.h"
+#include "WHashMap.h"
 
 /**
  * @brief      UrlEncoded Stream
@@ -45,7 +45,7 @@ public:
 	/**
 	 * @brief Return the total length of the stream
 	 * @retval int -1 is returned when the size cannot be determined
-	*/
+	 */
 	int available()
 	{
 		return stream.available();
@@ -93,4 +93,4 @@ private:
 };
 
 /** @} */
-#endif /* _SMING_CORE_DATA_UESTREAM_H_ */
+#endif /* _SMING_CORE_DATA_URL_ENCODDED_OUTPUT_STREAM_H_ */
