+++ Sming/SmingCore/Network/Http/HttpRequest.h
@@ -19,6 +19,7 @@
 #endif
 #include "../TcpConnection.h"
 #include "Data/Stream/DataSourceStream.h"
+#include "Data/Stream/MultipartStream.h"
 #include "Network/Http/HttpHeaders.h"
 
 class HttpClient;
