+++ Sming/SmingCore/Network/TcpClient.h
@@ -15,7 +15,7 @@
 #define _SMING_CORE_TCPCLIENT_H_
 
 #include "TcpConnection.h"
-#include "../Delegate.h"
+#include "Delegate.h"
 
 #ifdef ENABLE_SSL
 #include "SslValidator.h"
